3-11

+++ 6320500697_3.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main()
 {
-    int N,j,k;
+    long int N,j,k;
     scanf ("%d",&N);
-    j= N%3;
-    k= N%11;
+    j=N%3;
+    k=N%11;
     printf ("%d %d",j,k);
     return 0;
 }
